Add `modifying_machine_id`
This will allow us to click a button to set which machine should be
changing status, trying to reserve, etc.

It uses the label and then the resulting ID from the labe. There is
potentially other ways to do this as well, including passing the machine
to the button that it corresponds to.